+++ CodeWars/Algorithms/VasyaClerk/Java/src/LineTest.java
@@ -1,0 +1,19 @@
+/**
+ * Created by cajetan on 7/15/17.
+ */
+
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+import org.junit.runners.JUnit4;
+
+public class LineTest {
+    @Test
+    public void test1() {
+        assertEquals("YES", Line.Tickets(new int[] {25, 25, 50}));
+    }
+
+    @Test
+    public void test2() {
+        assertEquals("NO", Line.Tickets(new int []{25, 100}));
+    }
+}